New structured added to keep it organized
